+++ R/redist_ms_parallel.R
@@ -103,7 +103,7 @@ redist_mergesplit_parallel = function(map, nsims, chains=1,
         init_plans = get_plans_matrix(
             redist_smc(map, chains, counties, compactness, constraints,
                        resample=TRUE, adapt_k_thresh=adapt_k_thresh,
-                       ref_name=FALSE, verbose=verbose, silent=silent, cores=1))
+                       ref_name=FALSE, verbose=verbose, silent=silent, ncores=1))
         if (is.null(init_name))
             init_names = paste0("<init> ", seq_len(chains))
         else
@@ -189,7 +189,7 @@ redist_mergesplit_parallel = function(map, nsims, chains=1,
     doParallel::registerDoParallel(cl)
     on.exit(stopCluster(cl))
 
-    out_par <- foreach(chain=seq_len(chains), .inorder=FALSE) %dopar% {
+    out_par <- foreach(chain=seq_len(chains), .inorder=FALSE) %dorng% {
         if (!silent) cat("Starting chain ", chain, "\n", sep="")
         run_verbosity = if (chain == 1 || verbosity == 3) verbosity else 0
         ms_plans(nsims, adj, init_plans[, chain], counties, pop,