+++ explore/ex_nc.R
@@ -4,7 +4,7 @@ library(alarmdata)
 
 nc = alarm_50state_map("NC")
 
-#plans = redist_smc(nc, 200, counties=county, runs=4, cores=2, verbose=TRUE)
+#plans = redist_smc(nc, 200, counties=county, runs=4, ncores=2, verbose=TRUE)
 plans = redist_mergesplit_parallel(nc, 300, counties=county, chains=4, verbose=TRUE)
 
 plans = plans %>%
@@ -15,3 +15,5 @@ plans = plans %>%
            dev = plan_parity(nc))
 
 summary(plans)
+
+